FIX strlcat

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: sserbin <stanleyserbin@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/19 16:04:33 by sserbin           #+#    #+#             */
-/*   Updated: 2020/11/19 16:37:37 by sserbin          ###   ########.fr       */
+/*   Updated: 2020/12/06 22:21:16 by sserbin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ size_t	ft_strlcat(char *dest, char *str, size_t x)
 {
 	size_t	i;
 	int		n;
-	int		z;
+	size_t	z;
 
 	i = 0;
 	while (dest[i] && i < x)
